+++ src/main/kotlin/me/nobaboy/nobaaddons/ui/TextHudElement.kt
@@ -9,20 +9,33 @@ import net.minecraft.text.Text
 import kotlin.collections.maxOf
 import kotlin.collections.plus
 
-abstract class TextHud(element: AbstractTextElement<*>) : HudElement(element.position) {
+/**
+ * Generic text display [HudElement], providing helper methods to render text
+ * on screen.
+ */
+abstract class TextHudElement(element: AbstractTextElement<*>) : HudElement(element.position) {
 	open val color: Int by element::color
 	open val textShadow: TextShadow by element::textShadow
 	open val outlineColor: Int by element::outlineColor
 
+	/**
+	 * Calls [renderLine] for each of the provided [lines]
+	 */
 	protected fun renderLines(context: DrawContext, lines: List<Text>, x: Int = 0, y: Int = 0) {
 		lines.forEachIndexed { i, line -> renderLine(context, line, x, y, i) }
 	}
 
+	/**
+	 * Renders the provided [text] on the given line
+	 *
+	 * @param x How much to offset the X value for this line
+	 * @param y How much to offset the Y value for this line
+	 * @param line Automatically adds `line * fontRenderer.fontHeight` to [y]
+	 */
 	protected fun renderLine(context: DrawContext, text: Text, x: Int = 0, y: Int = 0, line: Int = 0) {
-		var x = this.x + x
-		if(alignment == ElementAlignment.RIGHT) {
-			val size = this.size
-			x += size.first - text.getWidth()
+		require(line >= 0) { "line must not be negative" }
+		val x = this.x + x + let {
+			if(alignment == ElementAlignment.RIGHT) this.size.first - text.getWidth() else 0
 		}
 		val y = this.y + y + (line * MCUtils.textRenderer.fontHeight)
 		when(textShadow) {
@@ -34,14 +47,19 @@ abstract class TextHud(element: AbstractTextElement<*>) : HudElement(element.pos
 		}
 	}
 
+	/**
+	 * Implement this method to render the text for this HUD
+	 *
+	 * This method is called from within [RenderUtils.scaled]
+	 */
 	protected abstract fun renderText(context: DrawContext)
 
 	override fun render(context: DrawContext) {
 		if(!shouldRender()) return
 
-		RenderUtils.startScale(context, scale)
-		renderText(context)
-		RenderUtils.endScale(context)
+		RenderUtils.scaled(context, scale) {
+			renderText(context)
+		}
 	}
 
 	/**