+++ src/main/kotlin/me/nobaboy/nobaaddons/ui/data/ElementBounds.kt
@@ -1,5 +1,8 @@
 package me.nobaboy.nobaaddons.ui.data
 
+/**
+ * Element bounds, primarily used by the HUD editor screen to restrict the element from being moved off the screen
+ */
 data class ElementBounds(
 	val x: Int,
 	val y: Int,