+++ src/main/kotlin/me/nobaboy/nobaaddons/ui/data/TextElement.kt
@@ -3,11 +3,19 @@ package me.nobaboy.nobaaddons.ui.data
 import com.google.gson.JsonObject
 import me.nobaboy.nobaaddons.ui.TextShadow
 
+/**
+ * Generic implementation of an [AbstractTextElement], suitable for basic text elements that don't require
+ * any extra properties.
+ *
+ * ## Example
+ *
+ * ```kt
+ * object Config : AbstractConfig(...) {
+ *     val element by TextElement("element", color = 0xFF00FF)
+ * }
+ * ```
+ */
 class TextElement(key: String = "") : AbstractTextElement<TextElement>(key) {
-	constructor(json: JsonObject) : this() {
-		load(json)
-	}
-
 	constructor(
 		key: String = "",
 		textShadow: TextShadow? = null,
@@ -23,6 +31,10 @@ class TextElement(key: String = "") : AbstractTextElement<TextElement>(key) {
 			this.position.y = it.y
 			this.position.scale = it.scale
 		}
+	}
+
+	constructor(json: JsonObject) : this() {
+		load(json)
 	}
 
 	fun copy() = TextElement(save() as JsonObject)